🚚 #7 元のページを削除して、リストページを日記トップに移動
